# ENH: optimize min max and median scipy.stats.binned_statistic