@lutefiskhotdish 2021-08-21
Would it be worth doing more than one benchmark with samples of different sizes and different numbers of bins?  I think in theory sorting (proposed enhancement) takes worst-case time `O(N log N)`, whereas applying `min()` or `max()` to each bin separately (existing numpy) would be `d * O(N)`, where `d` is the number of bins.  If that's right, and *if the number of bins is fixed*, then for very large samples the new approach should be slower than the existing approach.  Does that make sense?