@teddyort 2021-08-21
Thanks @lutefiskhotdish for your comment! Your point is clear, concise, and, I believe, theoretically correct.

However, I think the proposed algorithm will outperform the current one even when `d << N`. The reason is that the cause of the slow runtime of the current implementation is not just the `d` but in the non-vectorized code in the `_calc_binned_statistic` and `_create_binned_data` functions that get called even when `d=1`.  In particular, those functions call `np.unique` (which is also `O(N log N)` ) as well as some other loops over the bins which causes the bulk of the slow runtime. 

See [here](https://stackoverflow.com/a/52208910/1139661) for an explanation of why big-O is not always the most helpful way to analyze numpy performance.

The current benchmark seems to test using a fixed number of samples (3333) and a fixed number of bins (200x200)=40000. I agree it would be wise to add some other permutations, but I'm not sure the procedure to add benchmarks or change their parameters. In particular, it might make it hard to compare the before/after if we change the benchmark here which is why I have left it alone in this PR.

Do you have any thoughts on what would be a good new benchmark to add?