@lutefiskhotdish 2021-08-22
@teddyort that's interesting.  

So `_calc_binned_statistic` is O(N log N) because it calls `np.unique(bin_numbers)` and `np.unique` sorts.  Theoretically, unique values could be obtained in O(N) by hashing the values, and I note that `_create_binned_data` hashes the unique value anyway, putting them in a `dict` `bin_map`.

Does this mean it might be feasible to achieve O(N) performance for some binned statistics, like `count`, `min`, `max`?