@lutefiskhotdish 2021-08-22
> Do you have any thoughts on what would be a good new benchmark to add?

@teddyort For benchmarks I guess it would be helpful to get a sense of how the runtime increases with the number of samples (especially when there are at least some bins with many samples), and how it increases with the number of bins.  So, maybe some 1D data with 3000 samples versus 30000 samples from a normal distribution with 20 bins?  Then, maybe augment the existing 2D data (3333 uniform samples with 200x200 bins) with bivariate normal data (same number of samples, but concentrated mostly in a few bins), and maybe the same again but with 20x20 bins?