@teddyort 2021-08-22
> Does this mean it might be feasible to achieve O(N) performance for some binned statistics, like count, min, max?

Agreed! I'm pretty sure in theory it could be done for all of them. I may take a stab at it at some point but I'm not sure when I'll have the time. For now, this PR focuses on just the vectorization improvement since avoiding the call to `_create_binned_data` saves a lot of the time already.