@lutefiskhotdish 2021-08-29
For what it's worth, I tried this kind of thing:
```
...
    elif statistic == 'min':
        result.fill(np.nan)
        for vv in builtins.range(Vdim):
            _binned_min(binnumbers, values[vv], result[vv])
...
```
with
```
def _binned_min(bins, values, result):
    result[bins] = np.inf
    for bnum, val in zip(bins, values):
        if val < result[bnum]:
            result[bnum] = val
```
This runs faster than the existing implementation but not as fast as the proposed vectorized version, with samples up to a few million where the vectorized version takes about 1.6 seconds to run and the existing implementation takes about 5 seconds.  The `_binned_min` approach should eventually out-perform the vectorized version (because of big-O considerations), but if that doesn't happen until the samples exceed 100 million it hardly seems worthwhile.