@teddyort 2021-09-02
@lutefiskhotdish Thanks for sharing!

It would be very interesting to see if finding these stats could be done more efficiently (as it theoretically should). I think the key will be finding ways to avoid sorting and _also_ use vectorization to get the best of both worlds.

Still, since the proposed method substantially improves the current benchmarks, and probably most reasonable ones we could add, I would advocate merging it as is.

I'd be interested in looking for even better methods (and happy to collaborate) in the future, but not sure when I'll have the time.