@rgommers 2021-09-06
> Theoretically, unique values could be obtained in O(N) by hashing the values, and I note that `_create_binned_data` hashes the unique values anyway, putting them in a `dict` `bin_map`.

There have been suggestions of adding an unsorted option to `numpy.unique` so it could do exactly this (e.g., https://github.com/numpy/numpy/issues/11136, also some other discussions I cannot find back right now). That would be a nice enhancement that could come in handy in multiple places.